Clean code cluster implementation
